+++ packages/react-core/src/index.d.ts
@@ -1,9 +1,24 @@
+export {
+  getRequest,
+  postRequest,
+  encodeParameter,
+  REQUEST_GET_MAX_URL_LENGTH
+} from './utils/requestsUtils';
+
+export { debounce } from './utils/debounce';
+export { throttle } from './utils/throttle';
+export { randomString } from './utils/randomString';
+
+export { makeClosedInterval } from './utils/makeClosedInterval';
+
 export { AggregationTypes } from './operations/aggregation/AggregationTypes';
 export { aggregationFunctions } from './operations/aggregation/values';
 export { groupValuesByColumn } from './operations/groupby';
 export { histogram } from './operations/histogram';
+export { scatterPlot } from './operations/scatterPlot';
 
-export { viewportFeatures as viewportFeaturesBinary } from './filters/viewportFeaturesBinary';
+export { viewportFeatures } from './filters/viewportFeatures';
+export { viewportFeaturesBinary } from './filters/viewportFeaturesBinary';
 export { viewportFeaturesGeoJSON } from './filters/viewportFeaturesGeoJSON';
 
 export { AggregationFunctions, GroupByFeature, HistogramFeature, Viewport, ViewportFeaturesBinary } from './types';