+++ packages/react-ui/src/widgets/HistogramWidgetUI.js
@@ -260,7 +260,7 @@ function HistogramWidgetUI(props) {
         <Grid
           container
           direction='row'
-          justify='space-between'
+          justifyContent='space-between'
           alignItems='center'
           className={classes.optionsSelectedBar}
         >