+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Not released
 
+- Update deck.gl version to 8.5.0 [#162](https://github.com/CartoDB/carto-react/pull/162)
+- Fix Material-UI warnings on justify property [#162](https://github.com/CartoDB/carto-react/pull/162)
+
 ## 1.1.0-alpha.5 (2021-07-12)
 
 - Add and improve typings [#157](https://github.com/CartoDB/carto-react/pull/157)