+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Improve types [#160](https://github.com/CartoDB/carto-react/pull/160)
 - Update deck.gl version to 8.5.0 [#162](https://github.com/CartoDB/carto-react/pull/162)
 - Fix Material-UI warnings on justify property [#162](https://github.com/CartoDB/carto-react/pull/162)